ref(browser): Move normalizeStringifyValue to @sentry/browser
Relocates `normalizeStringifyValue` (and its test) out of
`@sentry-internal/browser-utils` into `@sentry/browser`, exports it from
the browser package's public surface, and switches `@sentry/vue` and
`@sentry/react` to import from `@sentry/browser` instead.

Both framework packages no longer need a direct dependency on
`@sentry-internal/browser-utils`, so that entry is dropped from their
`package.json` files.
