+++ packages/browser/test/normalizeStringifyValue.test.ts
