+++ packages/react/package.json
@@ -39,7 +39,6 @@
     "access": "public"
   },
   "dependencies": {
-    "@sentry-internal/browser-utils": "10.54.0",
     "@sentry/browser": "10.54.0",
     "@sentry/core": "10.54.0"
   },
