+++ packages/react/src/sdk.ts
@@ -1,8 +1,11 @@
 import type { BrowserOptions } from '@sentry/browser';
-import { init as browserInit, setContext } from '@sentry/browser';
+import {
+  init as browserInit,
+  normalizeStringifyValue as browserNormalizeStringifyValue,
+  setContext,
+} from '@sentry/browser';
 import type { Client } from '@sentry/core/browser';
 import { applySdkMetadata, setNormalizeStringifier } from '@sentry/core/browser';
-import { normalizeStringifyValue as browserNormalizeStringifyValue } from '@sentry-internal/browser-utils';
 import { version } from 'react';
 import { isSyntheticEvent } from './isSyntheticEvent';
